// frontend/src/api/models/acceptCategorizationRuleSuggestionResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { CategorizationRuleSuggestion } from './categorizationRuleSuggestion'

export interface AcceptCategorizationRuleSuggestionResponse {
  suggestion: CategorizationRuleSuggestion
}