// frontend/src/api/models/categorizationRuleRecommendationGenerationResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { CategorizationRuleSuggestionGeneration } from './categorizationRuleSuggestionGeneration'
import type { CategorizationRuleSuggestion } from './categorizationRuleSuggestion'

export interface CategorizationRuleRecommendationGenerationResponse {
  generation: CategorizationRuleSuggestionGeneration
  suggestions: CategorizationRuleSuggestion[]
}